requirements.txt
# Instagram Trend Reporter Dependencies
apify-client>=1.0.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
keyring>=23.0.0
pyyaml>=6.0.0

# Streamlit Web App
streamlit>=1.30.0

# Visualization
plotly>=5.18.0
matplotlib>=3.8.0
wordcloud>=1.9.0
kaleido>=0.2.1
